majority of styling completed
